Added a few more Fnooblatz1000 instructions, enough to draw pictures.
I added a little test script that draws an exclamation mark picture.
Also I put in some instructions so you can draw | and - and move
back and forth using the "alternator" (that changes each time a
Fnooblatz instruction is executed).  I haven't added a test yet for
those because I'm planning to test them with a pretty picture too! :D