Moved test runner so it runs every time the script runs as main.
Also I had to make sure the test runner leaves the Fnooblatz clean
and useful.  This simplifies my main use case for this script at
the moment, and makes it more certain the tests will be run regularly.
Theoretically they might have to be moved back out again if they
were to someday take a long time to run??  But that seems distant and
unlikely!