Added instruction to move to upper left corner.
I added a Fnooblatz 1000 instruction to move to the upper left corner,
and some random twiddler distributions to use it.  It doesn't seem to
be very useful right now!  But maybe if I add conditionals, since some
things output to the upper left corner, it'll be useful to go there to
check their values.