Added some new movement buttons.
I added buttons that allow you to move easily in the cardinal
directions.  (A much needed feature.)  The left-right buttons also
halt on a '|' and the up-down ones halt on a '-', so hopefully it will
help add some order to things.  A Fnooblatz1000 isn't quite
programmable yet, but at least its behavior does vary a little bit
depending on its past output!