First commit of the exciting new Fnooblatz **2000**!
The Fnooblatz 2000 basically allows me to make words that are
shortcuts for a series of buttons on the 1000, so that I can (a)
remember the commands more easily and (b) combine them into more
complicated ones.

The frame I made for playing with the Fnooblatz 2000 uses curses, and
helpfully highlights the cursor position.