Fnooblatz 2000 random twiddler.
I added a twiddler for the Fnooblatz 2000.  It does sequences and
distributions separately-- I'm not really sure why I didn't do it that
way at first with the 1000.  This time I also made it so the name of
the distribution in use is displayed at the bottom of the screen.